BAU Refactor Integration tests
