+++ test/utils/testdata/notifications/NotificationsTestData.scala
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2021 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package testdata.notifications
+
+import base.TestData._
+import models.{Notification, NotificationDetails}
+import reactivemongo.bson.BSONObjectID
+
+object NotificationsTestData {
+
+  val exampleParsedNotification =
+    Notification(BSONObjectID.generate(), payload, Some(NotificationDetails(fileReference, outcomeSuccess, filename)), dateTime)
+
+  val exampleUnparsedNotification = Notification(BSONObjectID.generate(), payload, None, dateTime)
+
+}
